+++ packages/ui/src/index.ts
@@ -75,10 +75,11 @@ export { DashboardPage, type DashboardPageProps, type MatchFocus } from './pages
 export { LpEditorPage } from './pages/LpEditorPage'
 export { MatchPage } from './pages/MatchPage'
 export { RankPage } from './pages/RankPage'
-export { SearchPage } from './pages/SearchPage'
+export { PlayersPage, type PlayerLink } from './pages/PlayersPage'
 export { RankInput, RankLabel, isApexTier } from './pages/RankInput'
 export { ServerDataPage, type ServerDataPageProps } from './admin/ServerData'
 export { InvitesPage, type InvitesPageProps } from './admin/Invites'
+export { LeagueAccountsPage, type LeagueAccountsPageProps } from './admin/LeagueAccounts'
 export { MembersPage, type MembersPageProps } from './admin/Members'
 export {
   ChangeEmailCard,